split out operations from state store implemenation for better code reuse
